Patch pagination count bug.
Laravel's Illuminate\Database package currently has a bug in it related to the
count of paginated results when a complex pagination query is created (multiple joins, etc), see https://github.com/laravel/framework/issues/19444

Holloway is built on top of the Illuminate\Database package and as such has this bug also.
I have patched it by ensuring that the $columns specified in the paginate() method are always passed down through to the getCountForPagination() method.